requirements.txt
django
djangorestframework
Pillow
django-cors-headers
python-decouple
authy
twilio
django-phonenumber-field[phonenumbers]